Relocate SimulationRunner module and name
